Add Azure Pipelines YAML file and baud rate patch
This uses a custom Docker container based on Debian Buster to patch
arm-trusted-firmware's baud rate to match U-Boot's default 1.5Mbps, build
arm-trusted-firmware and u-boot, and create the Rockchip loader files.

Builds separate binaries for env-in-MMC and env-in-SPI-flash. Also
creates an SD card image which can be used to write u-boot to SPI flash.

If building a tag v*-ci it will also publish the idbloader.img and u-boot.itb
artifacts to a Github Release.